refactor(admin): generateSemesterOptions awardHelpers로 이동, getSemesterTerm 재사용, useMemo 제거
