+++ frontend/src/utils/awardHelpers.ts
@@ -1,4 +1,4 @@
-import { Award, SemesterTerm } from '@/types/club';
+import { Award, SemesterTerm, SemesterTermType } from '@/types/club';
 
 export const formatSemesterLabel = (award: Award): string | null => {
   if (award?.year && award?.semesterTerm) {
@@ -15,4 +15,37 @@ export const getAwardKey = (award: Award, index: number): string =>
 export const getAwardSortValue = (award: Award): number => {
   const semesterValue = award.semesterTerm === SemesterTerm.FIRST ? 1 : 2;
   return award.year * 10 + semesterValue;
+};
+
+export interface SemesterOption {
+  year: number;
+  semesterTerm: SemesterTermType;
+  label: string;
+}
+
+export const generateSemesterOptions = (
+  existingAwards: Award[],
+): SemesterOption[] => {
+  const currentYear = new Date().getFullYear();
+  const options: SemesterOption[] = [];
+
+  for (let year = currentYear - 3; year <= currentYear + 1; year++) {
+    options.push({
+      year,
+      semesterTerm: SemesterTerm.FIRST,
+      label: `${year} 1학기`,
+    });
+    options.push({
+      year,
+      semesterTerm: SemesterTerm.SECOND,
+      label: `${year} 2학기`,
+    });
+  }
+
+  return options.filter(
+    (opt) =>
+      !existingAwards.some(
+        (a) => a.year === opt.year && a.semesterTerm === opt.semesterTerm,
+      ),
+  );
 };
